requirements.txt
streamlit>=1.28.0
pymupdf4llm>=0.0.5
lisette>=0.0.36,<0.1.0
litellm>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
